Changed directory structure for easier files manage in the future, plus minor fixes
