+++ django_api/etools_prp/apps/unicef/ppd_sync/item_get_partner.py
@@ -1,28 +1,26 @@
 import logging
+from typing import Optional
 
 from rest_framework.exceptions import ValidationError
 
 from etools_prp.apps.partner.models import Partner
 from etools_prp.apps.partner.serializers import PMPPartnerSerializer
+from etools_prp.apps.unicef.utils import process_model
 
 logger = logging.getLogger(__name__)
 
 
-def process_programme_item_get_partner(partner_org):
-    from etools_prp.apps.unicef.tasks import process_model
-
-    # Get partner data
-    partner_data = partner_org
+def update_create_partner(partner_data: dict) -> Optional[Partner]:
 
     # Skip entries without unicef_vendor_number
     if not partner_data['unicef_vendor_number']:
-        logger.warning("No unicef_vendor_number - skipping!")
-        return False
+        logger.warning(f"No unicef_vendor_number for PD - skipping!")
+        return None
 
     # Create/Assign Partner
     if not partner_data['name']:
-        logger.warning("No partner name - skipping!")
-        return False
+        logger.warning(f"No partner name for PD - skipping!")
+        return None
 
     partner_data['external_id'] = partner_data.get('id', '#')
 
@@ -37,5 +35,6 @@ def process_programme_item_get_partner(partner_org):
 
         return partner
     except ValidationError:
-        logger.exception('Error trying to save Partner model with {}'.format(partner_data))
-        return False
+        logger.exception('Error trying to save Partner model with {}'
+                         .format(partner_data))
+        return None
